requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
black==21.9b0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cycler==0.10.0
dj-database-url==0.5.0
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
future==0.18.2
gunicorn==20.1.0
idna==3.2
iso8601==0.1.16
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
mypy-extensions==0.4.3
numpy==1.21.2
pathspec==0.9.0
PeakUtils==1.3.3
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.1
pycparser==2.20
pyparsing==3.0.3
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.9.30
requests==2.26.0
scipy==1.7.1
serial==0.0.97
six==1.16.0
sounddevice==0.4.3
soupsieve==2.2.1
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
whitenoise==5.3.0
